RAP 3 Compatibility + Number format

+++ bundles/org.eclipse.rap.chartjs/src/org/eclipse/rap/chartjs/ChartPointData.java
@@ -17,27 +17,24 @@ import java.util.List;
 import org.eclipse.rap.json.JsonArray;
 import org.eclipse.rap.json.JsonObject;
 
-
 public class ChartPointData {
 
-  private final List<Integer> points = new ArrayList<Integer>( 5 );
-  private final List<ChartStyle> colors = new ArrayList<ChartStyle>( 5 );
+	private final List<Float> points = new ArrayList<Float>(5);
+	private final List<ChartStyle> colors = new ArrayList<ChartStyle>(5);
 
-  public ChartPointData addPoint( int value, ChartStyle chartStyle ) {
-    points.add( value );
-    colors.add( chartStyle );
-    return this;
-  }
+	public ChartPointData addPoint(float value, ChartStyle chartStyle) {
+		points.add(value);
+		colors.add(chartStyle);
+		return this;
+	}
 
-  JsonArray toJson() {
-    JsonArray result = new JsonArray();
-    for( int i = 0; i < points.size(); i++ ) {
-      result.add( new JsonObject()
-        .add( "value", points.get( i ) )
-        .add( "color", ChartStyle.asCss( colors.get( i ).getFillColor() ) )
-      );
-    }
-    return result;
-  }
+	JsonArray toJson() {
+		JsonArray result = new JsonArray();
+		for (int i = 0; i < points.size(); i++) {
+			result.add(new JsonObject().add("value", points.get(i)).add("color",
+					ChartStyle.asCss(colors.get(i).getFillColor())));
+		}
+		return result;
+	}
 
 }
